Re-implement document.getElementsByClassName
Write a function getElementsByClassName that returns all the nodes in
a DOM that have a certain class name. Use an inner function that makes
a recursive call to itself in order to check all elements in the DOM
for the class name. Return an array of the nodes with matching class
names.